+++ common/src/commonMain/kotlin/dev/inmo/micro_utils/common/Clamp.kt
@@ -1,10 +1,5 @@
 package dev.inmo.micro_utils.common
 
+@Deprecated("Redundant", ReplaceWith("coerceIn(min, max)"))
 @Suppress("NOTHING_TO_INLINE")
-inline fun <T : Comparable<T>> T.clamp(min: T, max: T): T {
-    return when {
-        this < min -> min
-        this > max -> max
-        else -> this
-    }
-}
+inline fun <T : Comparable<T>> T.clamp(min: T, max: T): T = coerceIn(min, max)